requirements.txt
asgiref==3.5.2
Django==4.1
django-markdown==0.8.4
django-taggit==3.0.0
importlib-metadata==4.12.0
Markdown==3.4.1
Pillow==9.2.0
sqlparse==0.4.2
zipp==3.8.1
